HBX-1683: Move class 'org.hibernate.tool.hbm2x.visitor.JavaTypeFromValueVisitor' to 'org.hibernate.tool.internal.export.pojo.JavaTypeFromValueVisitor'
